LayoutTests/inspector/elements/highlight-node-scaled.html

<html>
<head>

<style>

body {
    margin: 0;
}

iframe {
    position: absolute;
    left: 83px;
    top: 53px;
    width: 200px;
    height: 200px;
}

</style>

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function dumpInspectorHighlightRects()
{
    var rectNames = ["margin", "border", "padding", "content"];
    var rects = window.internals.inspectorHighlightRects(document);
    for (var i = 0; i < rects.length; i++)
    {
        var rectName = (i < rectNames.length ? rectNames[i] : "untitled");
        var rect = rects.item(i);
        var line = rectName + " rect is " + rect.width + " x " + rect.height + " at (" + rect.top + ", " + rect.left + ")<br/>";
        document.getElementById("console").innerHTML += line;
    }
}

function onload()
{
    if (window.internals)
        window.internals.settings.setPageScaleFactor(2, 0, 0);
    runTest();
}
    
function test()
{
    WebInspector.inspectorView.setCurrentPanel(WebInspector.panels.elements);
    WebInspector.panels.elements.treeOutline.addEventListener(WebInspector.ElementsTreeOutline.Events.SelectedNodeChanged, selectedNodeChanged, this);

    function selectedNodeChanged(event)
    {
        var node = event.data;
        if (node.getAttribute("id") == "div")
            RuntimeAgent.evaluate("dumpInspectorHighlightRects()", InspectorTest.completeTest);
    }
    InspectorTest.evaluateInConsole("inspect(iframeDivElement)");
}

</script>
</head>

<body onload="onload()">
<iframe src="resources/highlight-node-scaled-iframe.html"></iframe>
<div id="console"></div>
</body>
</html>
