LayoutTests/inspector/elements/highlight-node-scaled-expected.txt


margin rect is 166 x 188 at (154, 192)
border rect is 138 x 160 at (168, 206)
padding rect is 118 x 140 at (178, 216)
content rect is 66 x 88 at (204, 242)
